+++ apito-client/src/app/dialog/confirm-box/confirm-box.component.ts
@@ -1,6 +1,8 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { AuthService } from 'src/app/services/auth.service';
+import { PersonService } from 'src/app/services/person.service';
+import { CompanyService } from 'src/app/services/company.service';
 
 @Component({
   selector: 'app-confirm-box',
@@ -9,7 +11,10 @@ import { AuthService } from 'src/app/services/auth.service';
 })
 export class ConfirmBoxComponent implements OnInit {
   deleteTxt = ""
-  constructor(private authService: AuthService,
+  constructor(
+    private authService: AuthService,
+    private personService: PersonService,
+    private companyService: CompanyService,
     @Inject(MAT_DIALOG_DATA) public data: any,
     public dialogRef: MatDialogRef<ConfirmBoxComponent>) { }
 
@@ -19,10 +24,10 @@ export class ConfirmBoxComponent implements OnInit {
 
   confirmAction(){
     if(this.data.from == 'person') {
-      this.authService.deletePerson(this.data._id);
+      this.personService.deletePerson(this.data._id);
     } else {
       if(this.data.from == 'company') {
-        this.authService.deleteCompany(this.data._id);
+        this.companyService.deleteCompany(this.data._id);
       }
     }
   }
