+++ apito-client/src/app/dialog/person-add-edit/person-add-edit.component.html
@@ -27,7 +27,7 @@
             </mat-form-field>
             <mat-form-field class="example-full-width">
                 <mat-label>Birthday</mat-label>
-                <input type="date" matInput formControlName="birthday" id="birthday" placeholder="Birthday"/>
+                <input type="text" matInput (focus)="handleDate()" formControlName="birthday" id="birthday" placeholder="Birthday (MM/DD/YYYYY)"/>
             </mat-form-field>
             <mat-form-field class="example-full-width">
                 <mat-label>Address</mat-label>
