+++ apito-server/index.js
@@ -8,6 +8,8 @@ const dotenv = require("dotenv")
 dotenv.config({ path: './config/config.env' })
 const auth = require("./src/routes/auth");
 const canvas = require("./src/routes/canvas");
+const person = require("./src/routes/person");
+const company = require("./src/routes/company");
 
 const fileUpload = require("express-fileupload");
 const cookieParser = require("cookie-parser");
@@ -55,6 +57,9 @@ app.use(hpp())
 
 app.use("/api/v1/auth", auth)
 app.use("/api/v1/canvas", canvas)
+app.use("/api/v1/person", person)
+app.use("/api/v1/company", company)
+
 
 
 app.use(errrorHandler)
