+++ apito-server/src/models/Company.js
@@ -9,7 +9,7 @@ const CompanySchema = new mongoose.Schema({
     email: {
         type: String,
         required: [true, "Please add an email"],
-        unique: true,   
+        unique: false,   
         match: [
             /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/,
             "Please add a valid email"
