+++ apito-server/src/routes/auth.js
@@ -8,15 +8,7 @@ const {
     updateDetails, 
     updatePassword, 
     logout,
-    deactivateUser, 
-    addPerson, 
-    getPerson, 
-    updatePerson, 
-    deletePerson,
-    addCompany,
-    getCompany,
-    updateCompany,
-    deleteCompany
+    deactivateUser
 } = require('../controller/auth');
 const { protect } = require('../middleware/auth');
 const router = express.Router() 
@@ -38,31 +30,6 @@ router
     .put('/password', updatePassword)
 router
     .get('/user', protect, loggedInUser)
-
-router
-    .post('/addPerson', addPerson)
-
-router
-    .get('/getPerson/:user_id', getPerson)
-
-router
-    .put('/updatePerson', updatePerson)
-
-router
-    .delete('/deletePerson/:person_id', deletePerson)
-
-
-router
-    .post('/addCompany', addCompany)
-
-router
-    .get('/getCompany/:user_id', getCompany)
-
-router
-    .put('/updateCompany', updateCompany)
-
-router
-    .delete('/deleteCompany/:company_id', deleteCompany)
 
 router
     .put('/deactivate', deactivateUser)
